crypto/marmots_vs_multisignatures/requirements.txt
pycryptodome
fastecdsa